added rememberMe functionality.
if logged in successfully, each time you open the app, it will go directly to the fridge. If newly registered, the same result- you can close and open the app again. The only time it will ask for you to re enter your login information is if you go to profile and click logout, then your fridge is saved, but you must enter the correct credentials to be let back in